[Update] 统一授权资产页面，修改apiUpdateAttr函数名称
