+++ apps/assets/templates/assets/_asset_user_auth_view_modal.html
@@ -70,7 +70,7 @@ function showAuth() {
         var msg = "{% trans 'Get auth info error' %}";
         toastr.error(msg)
     };
-    APIUpdateAttr({
+    requestApi({
         url: url,
         method: "GET",
         success: success,
