+++ apps/authentication/templates/authentication/_mfa_confirm_modal.html
@@ -38,7 +38,7 @@ $(document).ready(function () {
     var error = function () {
         $("#mfa_error").addClass("text-danger").html(codeError);
     };
-    APIUpdateAttr({
+    requestApi({
         url: url,
         method: "POST",
         body: JSON.stringify(data),
