+++ apps/ops/templates/ops/command_execution_create.html
@@ -255,7 +255,7 @@ function execute() {
         }
     }
 
-    APIUpdateAttr({
+    requestApi({
         url: url,
         body: JSON.stringify(data),
         method: 'POST',
