+++ apps/perms/api/user_permission.py
@@ -2,25 +2,23 @@
 #
 import time
 import traceback
+from functools import reduce
 import uuid
-from hashlib import md5
-from django.core.cache import cache
-from django.conf import settings
 from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from rest_framework.views import APIView, Response
 from rest_framework.generics import (
     ListAPIView, get_object_or_404, RetrieveAPIView
 )
-from django.utils.translation import ugettext as _
 from rest_framework.pagination import LimitOffsetPagination
 
 from common.permissions import IsValidUser, IsOrgAdminOrAppUser
 from common.tree import TreeNodeSerializer
-from common.utils import get_logger, get_object_or_none
+from common.utils import get_logger
 from ..utils import (
     AssetPermissionUtil, ParserNode,
 )
+from .mixin import UserPermissionCacheMixin, GrantAssetsMixin, NodesWithUngroupMixin
 from .. import const
 from ..hands import User, Asset, Node, SystemUser, NodeSerializer
 from .. import serializers
@@ -35,153 +33,6 @@ __all__ = [
     'ValidateUserAssetPermissionApi', 'UserGrantedNodesAsTreeApi',
     'UserGrantedNodesWithAssetsAsTreeApi', 'GetUserAssetPermissionActionsApi',
 ]
-
-
-class UserPermissionCacheMixin:
-    cache_policy = '0'
-    RESP_CACHE_KEY = '_PERMISSION_RESPONSE_CACHE_V2_{}'
-    CACHE_TIME = settings.ASSETS_PERM_CACHE_TIME
-    _object = None
-
-    def get_object(self):
-        return None
-
-    # 内部使用可控制缓存
-    def _get_object(self):
-        if not self._object:
-            self._object = self.get_object()
-        return self._object
-
-    def get_object_id(self):
-        obj = self._get_object()
-        if obj:
-            return str(obj.id)
-        return None
-
-    def get_request_md5(self):
-        path = self.request.path
-        query = {k: v for k, v in self.request.GET.items()}
-        query.pop("_", None)
-        query = "&".join(["{}={}".format(k, v) for k, v in query.items()])
-        full_path = "{}?{}".format(path, query)
-        return md5(full_path.encode()).hexdigest()
-
-    def get_meta_cache_id(self):
-        obj = self._get_object()
-        util = AssetPermissionUtil(obj, cache_policy=self.cache_policy)
-        meta_cache_id = util.cache_meta.get('id')
-        return meta_cache_id
-
-    def get_response_cache_id(self):
-        obj_id = self.get_object_id()
-        request_md5 = self.get_request_md5()
-        meta_cache_id = self.get_meta_cache_id()
-        resp_cache_id = '{}_{}_{}'.format(obj_id, request_md5, meta_cache_id)
-        return resp_cache_id
-
-    def get_response_from_cache(self):
-        # 没有数据缓冲
-        meta_cache_id = self.get_meta_cache_id()
-        if not meta_cache_id:
-            logger.debug("Not get meta id: {}".format(meta_cache_id))
-            return None
-        # 从响应缓冲里获取响应
-        key = self.get_response_key()
-        data = cache.get(key)
-        if not data:
-            logger.debug("Not get response from cache: {}".format(key))
-            return None
-        logger.debug("Get user permission from cache: {}".format(self.get_object()))
-        response = Response(data)
-        return response
-
-    def expire_response_cache(self):
-        obj_id = self.get_object_id()
-        expire_cache_id = '{}_{}'.format(obj_id, '*')
-        key = self.RESP_CACHE_KEY.format(expire_cache_id)
-        cache.delete_pattern(key)
-
-    def get_response_key(self):
-        resp_cache_id = self.get_response_cache_id()
-        key = self.RESP_CACHE_KEY.format(resp_cache_id)
-        return key
-
-    def set_response_to_cache(self, response):
-        key = self.get_response_key()
-        cache.set(key, response.data, self.CACHE_TIME)
-        logger.debug("Set response to cache: {}".format(key))
-
-    def get(self, request, *args, **kwargs):
-        self.cache_policy = request.GET.get('cache_policy', '0')
-
-        obj = self._get_object()
-        if obj is None:
-            logger.debug("Not get response from cache: obj is none")
-            return super().get(request, *args, **kwargs)
-
-        if AssetPermissionUtil.is_not_using_cache(self.cache_policy):
-            logger.debug("Not get resp from cache: {}".format(self.cache_policy))
-            return super().get(request, *args, **kwargs)
-        elif AssetPermissionUtil.is_refresh_cache(self.cache_policy):
-            logger.debug("Not get resp from cache: {}".format(self.cache_policy))
-            self.expire_response_cache()
-
-        logger.debug("Try get response from cache")
-        resp = self.get_response_from_cache()
-        if not resp:
-            resp = super().get(request, *args, **kwargs)
-            self.set_response_to_cache(resp)
-        return resp
-
-
-class GrantAssetsMixin:
-    serializer_class = serializers.AssetGrantedSerializer
-
-    def get_serializer(self, queryset, many=True):
-        assets_ids = []
-        system_users_ids = set()
-        for asset in queryset:
-            assets_ids.append(asset["id"])
-            system_users_ids.update(set(asset["system_users"]))
-        assets = Asset.objects.filter(id__in=assets_ids).only(
-            *self.serializer_class.Meta.only_fields
-        )
-        assets_map = {asset.id: asset for asset in assets}
-        system_users = SystemUser.objects.filter(id__in=system_users_ids).only(
-            *self.serializer_class.system_users_only_fields
-        )
-        system_users_map = {s.id: s for s in system_users}
-        data = []
-        for item in queryset:
-            i = item["id"]
-            asset = assets_map.get(i)
-            if not asset:
-                continue
-
-            _system_users = item["system_users"]
-            system_users_granted = []
-            for sid, action in _system_users.items():
-                system_user = system_users_map.get(sid)
-                if not system_user:
-                    continue
-                system_user.actions = action
-                system_users_granted.append(system_user)
-            asset.system_users_granted = system_users_granted
-            data.append(asset)
-        return super().get_serializer(data, many=True)
-
-    def search_queryset(self, assets):
-        search = self.request.query_params.get("search")
-        if not search:
-            return assets
-
-        assets_map = {asset['id']: asset for asset in assets}
-        assets_ids = set(assets_map.keys())
-        assets_ids_search = Asset.objects.filter(id__in=assets_ids).filter(
-            Q(hostname__icontains=search) | Q(ip__icontains=search)
-        ).values_list('id', flat=True)
-        assets_ids &= set(assets_ids_search)
-        return [assets_map.get(asset_id) for asset_id in assets_ids]
 
 
 class UserGrantedAssetsApi(UserPermissionCacheMixin, GrantAssetsMixin, ListAPIView):
@@ -203,7 +54,6 @@ class UserGrantedAssetsApi(UserPermissionCacheMixin, GrantAssetsMixin, ListAPIVi
         user = self.get_object()
         util = AssetPermissionUtil(user, cache_policy=self.cache_policy)
         queryset = util.get_assets()
-        queryset = self.search_queryset(queryset)
         return queryset
 
     def get_permissions(self):
@@ -212,29 +62,52 @@ class UserGrantedAssetsApi(UserPermissionCacheMixin, GrantAssetsMixin, ListAPIVi
         return super().get_permissions()
 
 
-class NodesWithUngroupMixin:
-    util = None
+class UserGrantedNodeAssetsApi(UserPermissionCacheMixin, GrantAssetsMixin, ListAPIView):
+    """
+    查询用户授权的节点下的资产的api, 与上面api不同的是，只返回某个节点下的资产
+    """
+    permission_classes = (IsOrgAdminOrAppUser,)
+    pagination_class = LimitOffsetPagination
 
-    @staticmethod
-    def get_ungrouped_node(ungroup_key):
-        return Node(key=ungroup_key, id=const.UNGROUPED_NODE_ID,
-                    value=_("ungrouped"))
+    def get_object(self):
+        user_id = self.kwargs.get('pk', '')
 
-    @staticmethod
-    def get_empty_node():
-        return Node(key=const.EMPTY_NODE_KEY, id=const.EMPTY_NODE_ID,
-                    value=_("empty"))
+        if user_id:
+            user = get_object_or_404(User, id=user_id)
+        else:
+            user = self.request.user
+        return user
 
-    def add_ungrouped_nodes(self, node_map, node_keys):
-        ungroup_key = '1:-1'
-        for key in node_keys:
-            if key.endswith('-1'):
-                ungroup_key = key
+    def get_node_key(self):
+        node_id = self.kwargs.get('node_id')
+        if str(node_id) == const.UNGROUPED_NODE_ID:
+            key = self.util.tree.ungrouped_key
+        elif str(node_id) == const.EMPTY_NODE_ID:
+            key = const.EMPTY_NODE_KEY
+        else:
+            node = get_object_or_404(Node, id=node_id)
+            key = node.key
+        return key
+
+    def get_queryset(self):
+        user = self.get_object()
+        self.util = AssetPermissionUtil(user, cache_policy=self.cache_policy)
+        key = self.get_node_key()
+        nodes_items = self.util.get_nodes_with_assets()
+        assets_system_users = {}
+        for item in nodes_items:
+            if item["key"] == key:
+                assets_system_users = item["assets"]
                 break
-        ungroup_node = self.get_ungrouped_node(ungroup_key)
-        empty_node = self.get_empty_node()
-        node_map[ungroup_key] = ungroup_node
-        node_map[const.EMPTY_NODE_KEY] = empty_node
+        assets = []
+        for asset_id, system_users in assets_system_users.items():
+            assets.append({"id": asset_id, "system_users": system_users})
+        return assets
+
+    def get_permissions(self):
+        if self.kwargs.get('pk') is None:
+            self.permission_classes = (IsValidUser,)
+        return super().get_permissions()
 
 
 class UserGrantedNodesApi(UserPermissionCacheMixin, NodesWithUngroupMixin, ListAPIView):
@@ -435,55 +308,6 @@ class UserGrantedNodesWithAssetsAsTreeApi(UserGrantedNodesWithAssetsApi):
         return self.serializer_class(queryset, many=True)
 
 
-class UserGrantedNodeAssetsApi(UserPermissionCacheMixin, GrantAssetsMixin, ListAPIView):
-    """
-    查询用户授权的节点下的资产的api, 与上面api不同的是，只返回某个节点下的资产
-    """
-    permission_classes = (IsOrgAdminOrAppUser,)
-    pagination_class = LimitOffsetPagination
-
-    def get_object(self):
-        user_id = self.kwargs.get('pk', '')
-
-        if user_id:
-            user = get_object_or_404(User, id=user_id)
-        else:
-            user = self.request.user
-        return user
-
-    def get_node_key(self):
-        node_id = self.kwargs.get('node_id')
-        if str(node_id) == const.UNGROUPED_NODE_ID:
-            key = self.util.tree.ungrouped_key
-        elif str(node_id) == const.EMPTY_NODE_ID:
-            key = const.EMPTY_NODE_KEY
-        else:
-            node = get_object_or_404(Node, id=node_id)
-            key = node.key
-        return key
-
-    def get_queryset(self):
-        user = self.get_object()
-        self.util = AssetPermissionUtil(user, cache_policy=self.cache_policy)
-        key = self.get_node_key()
-        nodes_items = self.util.get_nodes_with_assets()
-        assets_system_users = {}
-        for item in nodes_items:
-            if item["key"] == key:
-                assets_system_users = item["assets"]
-                break
-        assets = []
-        for asset_id, system_users in assets_system_users.items():
-            assets.append({"id": asset_id, "system_users": system_users})
-        assets = self.search_queryset(assets)
-        return assets
-
-    def get_permissions(self):
-        if self.kwargs.get('pk') is None:
-            self.permission_classes = (IsValidUser,)
-        return super().get_permissions()
-
-
 class ValidateUserAssetPermissionApi(UserPermissionCacheMixin, APIView):
     permission_classes = (IsOrgAdminOrAppUser,)
     
@@ -522,16 +346,12 @@ class GetUserAssetPermissionActionsApi(UserPermissionCacheMixin, RetrieveAPIView
         system_id = self.request.query_params.get('system_user_id', '')
 
         user = get_object_or_404(User, id=user_id)
-        asset = get_object_or_404(Asset, id=asset_id)
-        su = get_object_or_404(SystemUser, id=system_id)
 
         util = AssetPermissionUtil(user, cache_policy=self.cache_policy)
-        granted_assets = util.get_assets()
-        granted_system_users = granted_assets.get(asset, {})
-
-        _object = {}
-        if su not in granted_system_users:
-            _object['actions'] = 0
-        else:
-            _object['actions'] = granted_system_users[su]
-        return _object
+        assets = util.get_assets()
+        actions = 0
+        for asset in assets:
+            if asset_id == asset["id"]:
+                actions = asset["system_users"].get(system_id, 0)
+                break
+        return {"actions": actions}
