+++ apps/perms/templates/perms/asset_permission_user.html
@@ -160,7 +160,7 @@ function addUsers(users) {
     var success = function(data) {
         location.reload();
     };
-    APIUpdateAttr({
+    requestApi({
         url: the_url,
         body: JSON.stringify(body),
         success: success
@@ -175,7 +175,7 @@ function removeUser(users) {
     var success = function(data) {
         location.reload();
     };
-    APIUpdateAttr({
+    requestApi({
         url: the_url,
         body: JSON.stringify(body),
         success: success
@@ -187,7 +187,7 @@ function updateGroup(groups) {
     var body = {
         user_groups: groups
     };
-    APIUpdateAttr({
+    requestApi({
         url: the_url,
         body: JSON.stringify(body)
     });
