+++ apps/perms/templates/perms/remote_app_permission_remote_app.html
@@ -120,7 +120,7 @@
             var success = function(data) {
                 location.reload();
             };
-            APIUpdateAttr({
+            requestApi({
                 url: the_url,
                 body: JSON.stringify(body),
                 success: success
@@ -134,7 +134,7 @@
             var success = function(data) {
                 location.reload();
             };
-            APIUpdateAttr({
+            requestApi({
                 url: the_url,
                 body: JSON.stringify(body),
                 success: success
